Registering UnitHealthStore in constructor instead of StoreConfig
